requirements.txt
asgiref==3.7.2
certifi==2024.6.2
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
Django==5.0.2
django-crispy-forms==2.1
django-formtools==2.5.1
gunicorn==22.0.0
idna==3.7
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.32.3
sqlparse==0.4.4
stripe==9.12.0
typing_extensions==4.9.0
urllib3==2.2.2
